// src/components/map/minemap.d.ts
  export interface Source {
    type: "vector" | "raster" | "raster-dem" | "geojson" | "image" | "video" | "canvas";
  }
// ...
  export class LngLat {
    lng: number;
    lat: number;

    constructor(lng: number, lat: number);

    /** Return a new LngLat object whose longitude is wrapped to the range (-180, 180). */
    wrap(): mapboxgl.LngLat;

    /** Return a LngLat as an array */
    toArray(): number[];

    /** Return a LngLat as a string */
    toString(): string;

    /** Returns the approximate distance between a pair of coordinates in meters
     * Uses the Haversine Formula (from R.W. Sinnott, "Virtues of the Haversine", Sky and Telescope, vol. 68, no. 2, 1984, p. 159) */
    distanceTo(lngLat: LngLat): number;

    toBounds(radius: number): LngLatBounds;

    static convert(input: LngLatLike): mapboxgl.LngLat;
  }

  /**
   * LngLatBounds
   */
  export class LngLatBounds {
    sw: LngLatLike;
    ne: LngLatLike;

    constructor(boundsLike?: [LngLatLike, LngLatLike] | [number, number, number, number]);
    constructor(sw: LngLatLike, ne: LngLatLike);

    setNorthEast(ne: LngLatLike): this;

    setSouthWest(sw: LngLatLike): this;

    /** Check if the point is within the bounding box. */
    contains(lnglat: LngLatLike): boolean;

    /** Extend the bounds to include a given LngLat or LngLatBounds. */
    extend(obj: mapboxgl.LngLatLike | mapboxgl.LngLatBoundsLike): this;

    /** Get the point equidistant from this box's corners */
    getCenter(): mapboxgl.LngLat;

    /** Get southwest corner */
    getSouthWest(): mapboxgl.LngLat;

    /** Get northeast corner */
    getNorthEast(): mapboxgl.LngLat;

    /** Get northwest corner */
    getNorthWest(): mapboxgl.LngLat;

    /** Get southeast corner */
    getSouthEast(): mapboxgl.LngLat;

    /** Get west edge longitude */
    getWest(): number;

    /** Get south edge latitude */
    getSouth(): number;

    /** Get east edge longitude */
    getEast(): number;

    /** Get north edge latitude */
    getNorth(): number;

    /** Returns a LngLatBounds as an array */
    toArray(): number[][];

    /** Return a LngLatBounds as a string */
    toString(): string;

    /** Returns a boolean */
    isEmpty(): boolean;

    /** Convert an array to a LngLatBounds object, or return an existing LngLatBounds object unchanged. */
    static convert(input: LngLatBoundsLike): mapboxgl.LngLatBounds;
  }

  /**
   * Point
   */
  // Todo: Pull out class to seperate definition for Module "point-geometry"
  export class Point {
    x: number;
    y: number;

    constructor(x: number, y: number);

    clone(): Point;

    add(p: Point): Point;

    sub(p: Point): Point;

    mult(k: number): Point;

    div(k: number): Point;

    rotate(a: number): Point;

    matMult(m: number): Point;

    unit(): Point;

    perp(): Point;

    round(): Point;

    mag(): number;

    equals(p: Point): boolean;

    dist(p: Point): number;

    distSqr(p: Point): number;

    angle(): number;

    angleTo(p: Point): number;

    angleWidth(p: Point): number;

    angleWithSep(x: number, y: number): number;

    static convert(a: PointLike): Point;
  }
// ...
  export type EventData = { [key: string]: any };

  export class MapboxEvent<TOrig = undefined> {
    type: string;
    target: Map;
    originalEvent: TOrig;
  }

  export class MapMouseEvent extends MapboxEvent<MouseEvent> {
    type:
      | "mousedown"
      | "mouseup"
      | "click"
      | "dblclick"
      | "mousemove"
      | "mouseover"
      | "mouseenter"
      | "mouseleave"
      | "mouseout"
      | "contextmenu";

    point: Point;
    lngLat: LngLat;

    preventDefault(): void;
    defaultPrevented: boolean;
  }

  export type MapLayerMouseEvent = MapMouseEvent & { features?: MapboxGeoJSONFeature[] };

  export class MapTouchEvent extends MapboxEvent<TouchEvent> {
    type: "touchstart" | "touchend" | "touchcancel";

    point: Point;
    lngLat: LngLat;
    points: Point[];
    lngLats: LngLat[];

    preventDefault(): void;
    defaultPrevented: boolean;
  }

  export type MapLayerTouchEvent = MapTouchEvent & { features?: MapboxGeoJSONFeature[] };

  export class MapWheelEvent extends MapboxEvent<WheelEvent> {
    type: "wheel";

    preventDefault(): void;
    defaultPrevented: boolean;
  }

  export interface MapBoxZoomEvent extends MapboxEvent<MouseEvent> {
    type: "boxzoomstart" | "boxzoomend" | "boxzoomcancel";

    boxZoomBounds: LngLatBounds;
  }

  export type MapDataEvent = MapSourceDataEvent | MapStyleDataEvent;

  export interface MapStyleDataEvent extends MapboxEvent {
    dataType: "style";
  }

  export interface MapSourceDataEvent extends MapboxEvent {
    dataType: "source";
    isSourceLoaded: boolean;
    source: Source;
    sourceId: string;
    sourceDataType: "metadata" | "content";
    tile: any;
    coord: Coordinate;
  }

  export interface Coordinate {
    canonical: CanonicalCoordinate;
    wrap: number;
    key: number;
  }

  export interface CanonicalCoordinate {
    x: number;
    y: number;
    z: number;
    key: number;
    equals(coord: CanonicalCoordinate): boolean;
  }

  export interface MapContextEvent extends MapboxEvent<WebGLContextEvent> {
    type: "webglcontextlost" | "webglcontextrestored";
  }

  export class ErrorEvent extends MapboxEvent {
    type: "error";
    error: Error;
  }
// ...
  export type MapEventType = {
    error: ErrorEvent;

    load: MapboxEvent;
    idle: MapboxEvent;
    remove: MapboxEvent;
    render: MapboxEvent;
    resize: MapboxEvent;

    webglcontextlost: MapContextEvent;
    webglcontextrestored: MapContextEvent;

    dataloading: MapDataEvent;
    data: MapDataEvent;
    tiledataloading: MapDataEvent;
    sourcedataloading: MapSourceDataEvent;
    styledataloading: MapStyleDataEvent;
    sourcedata: MapSourceDataEvent;
    styledata: MapStyleDataEvent;

    boxzoomcancel: MapBoxZoomEvent;
    boxzoomstart: MapBoxZoomEvent;
    boxzoomend: MapBoxZoomEvent;

    touchcancel: MapTouchEvent;
    touchmove: MapTouchEvent;
    touchend: MapTouchEvent;
    touchstart: MapTouchEvent;

    click: MapMouseEvent;
    contextmenu: MapMouseEvent;
    dblclick: MapMouseEvent;
    mousemove: MapMouseEvent;
    mouseup: MapMouseEvent;
    mousedown: MapMouseEvent;
    mouseout: MapMouseEvent;
    mouseover: MapMouseEvent;

    movestart: MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined>;
    move: MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined>;
    moveend: MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined>;

    zoomstart: MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined>;
    zoom: MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined>;
    zoomend: MapboxEvent<MouseEvent | TouchEvent | WheelEvent | undefined>;

    rotatestart: MapboxEvent<MouseEvent | TouchEvent | undefined>;
    rotate: MapboxEvent<MouseEvent | TouchEvent | undefined>;
    rotateend: MapboxEvent<MouseEvent | TouchEvent | undefined>;

    dragstart: MapboxEvent<MouseEvent | TouchEvent | undefined>;
    drag: MapboxEvent<MouseEvent | TouchEvent | undefined>;
    dragend: MapboxEvent<MouseEvent | TouchEvent | undefined>;

    pitchstart: MapboxEvent<MouseEvent | TouchEvent | undefined>;
    pitch: MapboxEvent<MouseEvent | TouchEvent | undefined>;
    pitchend: MapboxEvent<MouseEvent | TouchEvent | undefined>;

    wheel: MapWheelEvent;
  };